+++ network/depcheck_test.go
@@ -26,5 +26,6 @@ func TestNoDeps(t *testing.T) {
 	depcheck.AssertNoDependency(t, map[string][]string{
 		"knative.dev/pkg/network":          depcheck.KnownHeavyDependencies,
 		"knative.dev/pkg/network/handlers": depcheck.KnownHeavyDependencies,
+		"knative.dev/pkg/network/tls":      depcheck.KnownHeavyDependencies,
 	})
 }
